change the test behaviour
