+++ test/suite/utils/getNonce.test.ts
@@ -1,0 +1,29 @@
+/*
+ * Copyright (c) 1998-2025 KX Systems Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the
+ * License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
+
+import * as assert from "assert";
+
+import { getNonce } from "../../../src/utils/getNonce";
+
+describe("getNonce", () => {
+  it("should return a string with length 32", () => {
+    const nonce = getNonce();
+    assert.strictEqual(nonce.length, 32);
+  });
+
+  it("should return a string containing only alphanumeric characters", () => {
+    const nonce = getNonce();
+    const alphanumericRegex = /^[a-zA-Z0-9]+$/;
+    assert.match(nonce, alphanumericRegex);
+  });
+});
