+++ test/suite/utils/shell.test.ts
@@ -1,0 +1,31 @@
+/*
+ * Copyright (c) 1998-2025 KX Systems Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the
+ * License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
+
+import * as sinon from "sinon";
+
+import * as shellUtils from "../../../src/utils/shell";
+
+describe("shell", () => {
+  describe("killPid", () => {
+    let tryExecuteCommandStub: sinon.SinonStub;
+
+    beforeEach(() => {
+      tryExecuteCommandStub = sinon.stub();
+    });
+
+    it("should not execute command if pid is NaN", async () => {
+      await shellUtils.killPid(NaN);
+      sinon.assert.notCalled(tryExecuteCommandStub);
+    });
+  });
+});
